move test to projectfile
